+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Upcoming
 
 ### Breaking changes
 
+* Tx author needs to have an associated registered user to operate on Orgs
+* `Org::members` is now `Vec<Id>`
 * cli: Rename the key-pair storage file from 'accounts.json' to 'key-pairs.json'
 * cli: Move key-pair related commands under the new `key-pair` command group
 * client: `TransactionApplied` result is now `Result<(), TransactionError>`